add desktop menu
