+++ private/components/molecules/DesktopMenu/index.stories.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { storiesOf } from '@storybook/react';
+import Menu from '.';
+
+storiesOf('Menu', module)
+  .add('Desktop Menu', () => <Menu />);
